refactor: Playing과 Bikjang의 move() 중복 분기처리를 템플릿 메서드 패턴을 적용
