+++ src/main/java/domain/state/Bikjang.java
@@ -2,7 +2,6 @@ package domain.state;
 
 import domain.Board;
 import domain.piece.Team;
-import domain.position.Position;
 
 public class Bikjang extends Running {
     public Bikjang(long id, Board board, Team turn) {
@@ -10,14 +9,12 @@ public class Bikjang extends Running {
     }
 
     @Override
-    public JanggiGame move(Position start, Position destination) {
-        board.move(turn, start, destination);
-        if (board.isAnyJangDead()) {
-            return new Finished(id, board, turn.changeTeam());
-        }
-        if (board.isBikjang()) {
-            return new Finished(id, board, turn.changeTeam());
-        }
+    protected JanggiGame transitionOnBikjang() {
+        return new Finished(id, board, turn.changeTeam());
+    }
+
+    @Override
+    protected JanggiGame transitionOnNormal() {
         return new Playing(id, board, turn.changeTeam());
     }
